openenv_pkg/models.py
"""Data models for the EnterpriseSim Customer Support Environment."""

from openenv.core.env_server.mcp_types import CallToolAction, ListToolsAction
from server.environment import SupportObservation

# Re-export for OpenEnv convention
__all__ = ["CallToolAction", "ListToolsAction", "SupportObservation"]
