openenv_pkg/pyproject.toml
[project]
name = "enterprise-sim-openenv"
version = "0.1.0"
description = "OpenEnv RL environment for enterprise customer support training"
requires-python = ">=3.11"
dependencies = [
    "openenv-core[core]>=0.2.1",
    "fastapi>=0.115.0",
    "fastmcp",
    "pydantic>=2.0.0",
    "uvicorn>=0.24.0",
    "openai>=1.0",
    "requests>=2.31.0",
]

[build-system]
requires = ["setuptools>=45", "wheel"]
build-backend = "setuptools.build_meta"

[project.scripts]
server = "server.app:main"

[tool.setuptools]
include-package-data = true
packages = ["server"]
